+++ eslint.ignores.js
@@ -1,1 +1,14 @@
-export default ['build/'];
+export default [
+  'build/',
+  'dist/',
+  'node_modules/',
+  '.git/',
+  'coverage/',
+  '.eslintcache',
+  '*.config.js',
+  '*.config.ts',
+  'eslint.ignores.js',
+  'scripts/',
+  'pnpm-lock.yaml',
+  'public/',
+];